% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmice.R
\name{pmice}
\alias{pmice}
\title{Impute missing data in parallel}
\usage{
pmice(data, m = 5, method = vector("character", length = ncol(data)),
  predictorMatrix = (1 - diag(1, ncol(data))), where = is.na(data),
  visitSequence = NULL, form = vector("character", length = ncol(data)),
  post = vector("character", length = ncol(data)), defaultMethod = c("pmm",
  "logreg", "polyreg", "polr"), maxit = 5, diagnostics = TRUE,
  printFlag = FALSE, seed = NA, imputationMethod = NULL,
  defaultImputationMethod = NULL, data.init = NULL, ...)
}
\description{
Impute missing data in parallel
}
